Refactor(analysis): extract PcmLqnsAnalyzerContext from PcmLqnsAnalyzer
